fix: Use float mapping for fractional service/agent monitoring fields
ResponseTime, CpuPercentage and CycleDuration are floats (seconds and
percent) but were mapped as long, so OpenSearch truncated them in the
indexed values. Aggregations (avg, percentiles) therefore returned 0 for
all sub-second response times, while _source kept the correct values.

The new mapping only applies to newly created (weekly) indices; existing
indices need a reindex to recover correct aggregations.

+++ src/DIRAC/MonitoringSystem/Client/Types/ServiceMonitoring.py
@@ -40,14 +40,14 @@ class ServiceMonitoring(BaseType):
                 "Status": {"type": "keyword"},
                 "Location": {"type": "keyword"},
                 "MemoryUsage": {"type": "long"},
-                "CpuPercentage": {"type": "long"},
+                "CpuPercentage": {"type": "float"},
                 "Connections": {"type": "long"},
                 "Queries": {"type": "long"},
                 "PendingQueries": {"type": "long"},
                 "ActiveQueries": {"type": "long"},
                 "RunningThreads": {"type": "long"},
                 "MaxFD": {"type": "long"},
-                "ResponseTime": {"type": "long"},
+                "ResponseTime": {"type": "float"},
             }
         )
